Criação de Calculadora

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -34,7 +34,7 @@ public class Calculadora {
 	public static int subtracao(int a, int b) {
 		return a - b;
 	}
-	public static int divisao(int a, int b) {
+	public static double divisao(int a, int b) {
 		return a / b;
 	}
 	public static int multiplicacao(int a, int b) {
